+++ python/packages/core/agent_framework/_workflows/__init__.py
@@ -1,6 +1,11 @@
 # Copyright (c) Microsoft. All rights reserved.
 
 from ._agent import WorkflowAgent
+from ._agent_executor import (
+    AgentExecutor,
+    AgentExecutorRequest,
+    AgentExecutorResponse,
+)
 from ._checkpoint import (
     CheckpointStorage,
     FileCheckpointStorage,
@@ -42,9 +47,6 @@ from ._events import (
     WorkflowStatusEvent,
 )
 from ._executor import (
-    AgentExecutor,
-    AgentExecutorRequest,
-    AgentExecutorResponse,
     Executor,
     RequestInfoExecutor,
     RequestInfoMessage,
