+++ python/packages/core/agent_framework/_workflows/_executor.py
@@ -12,14 +12,9 @@ from dataclasses import asdict, dataclass, field, fields, is_dataclass
 from textwrap import shorten
 from typing import Any, ClassVar, Generic, TypeVar, cast
 
-from .._agents import AgentProtocol
-from .._threads import AgentThread
-from .._types import AgentRunResponse, AgentRunResponseUpdate, ChatMessage
 from ..observability import create_processing_span
 from ._checkpoint import WorkflowCheckpoint
 from ._events import (
-    AgentRunEvent,
-    AgentRunUpdateEvent,
     ExecutorCompletedEvent,
     ExecutorFailedEvent,
     ExecutorInvokedEvent,
@@ -1345,163 +1340,3 @@ class RequestInfoExecutor(Executor):
 
 
 # endregion: Request Info Executor
-
-# region Agent Executor
-
-
-@dataclass
-class AgentExecutorRequest:
-    """A request to an agent executor.
-
-    Attributes:
-        messages: A list of chat messages to be processed by the agent.
-        should_respond: A flag indicating whether the agent should respond to the messages.
-            If False, the messages will be saved to the executor's cache but not sent to the agent.
-    """
-
-    messages: list[ChatMessage]
-    should_respond: bool = True
-
-
-@dataclass
-class AgentExecutorResponse:
-    """A response from an agent executor.
-
-    Attributes:
-        executor_id: The ID of the executor that generated the response.
-    agent_run_response: The underlying agent run response (unaltered from client).
-    full_conversation: The full conversation context (prior inputs + all assistant/tool outputs) that
-        should be used when chaining to another AgentExecutor. This prevents downstream agents losing
-        user prompts while keeping the emitted AgentRunEvent text faithful to the raw agent output.
-    """
-
-    executor_id: str
-    agent_run_response: AgentRunResponse
-    full_conversation: list[ChatMessage] | None = None
-
-
-class AgentExecutor(Executor):
-    """built-in executor that wraps an agent for handling messages."""
-
-    def __init__(
-        self,
-        agent: AgentProtocol,
-        *,
-        agent_thread: AgentThread | None = None,
-        streaming: bool = False,
-        id: str | None = None,
-    ):
-        """Initialize the executor with a unique identifier.
-
-        Args:
-            agent: The agent to be wrapped by this executor.
-            agent_thread: The thread to use for running the agent. If None, a new thread will be created.
-            streaming: Enable streaming (emits incremental AgentRunUpdateEvent events) vs single response.
-            id: A unique identifier for the executor. If None, a new UUID will be generated.
-        """
-        # Prefer provided id; else use agent.name if present; else generate deterministic prefix
-        if id is not None:
-            exec_id = id
-        else:
-            agent_name = agent.name
-            if agent_name:
-                exec_id = str(agent_name)
-            else:
-                logger.warning("Agent has no name, using fallback ID 'executor_unnamed'")
-                exec_id = "executor_unnamed"
-        super().__init__(exec_id)
-        self._agent = agent
-        self._agent_thread = agent_thread or self._agent.get_new_thread()
-        self._streaming = streaming
-        self._cache: list[ChatMessage] = []
-
-    async def _run_agent_and_emit(self, ctx: WorkflowContext[AgentExecutorResponse]) -> None:
-        """Execute the underlying agent, emit events, and enqueue response.
-
-        Terminal detection is handled centrally in Runner.
-        This method only produces AgentRunEvent/AgentRunUpdateEvent plus enqueues an
-        AgentExecutorResponse message for routing.
-        """
-        if self._streaming:
-            updates: list[AgentRunResponseUpdate] = []
-            async for update in self._agent.run_stream(
-                self._cache,
-                thread=self._agent_thread,
-            ):
-                # Skip empty updates (no textual or structural content)
-                if not update:
-                    continue
-                contents = getattr(update, "contents", None)
-                text_val = getattr(update, "text", "")
-                has_text_content = False
-                if contents:
-                    for c in contents:
-                        if getattr(c, "text", None):
-                            has_text_content = True
-                            break
-                if not (text_val or has_text_content):
-                    continue
-                updates.append(update)
-                await ctx.add_event(AgentRunUpdateEvent(self.id, update))
-            response = AgentRunResponse.from_agent_run_response_updates(updates)
-        else:
-            response = await self._agent.run(
-                self._cache,
-                thread=self._agent_thread,
-            )
-            await ctx.add_event(AgentRunEvent(self.id, response))
-
-        # Always construct a full conversation snapshot from inputs (cache)
-        # plus agent outputs (agent_run_response.messages). Do not mutate
-        # response.messages so AgentRunEvent remains faithful to the raw output.
-        full_conversation: list[ChatMessage] = list(self._cache) + list(response.messages)
-
-        agent_response = AgentExecutorResponse(self.id, response, full_conversation=full_conversation)
-        await ctx.send_message(agent_response)
-        self._cache.clear()
-
-    @handler
-    async def run(self, request: AgentExecutorRequest, ctx: WorkflowContext[AgentExecutorResponse]) -> None:
-        """Handle an AgentExecutorRequest (canonical input).
-
-        This is the standard path: extend cache with provided messages; if should_respond
-        run the agent and emit an AgentExecutorResponse downstream.
-        """
-        self._cache.extend(request.messages)
-        if request.should_respond:
-            await self._run_agent_and_emit(ctx)
-
-    @handler
-    async def from_response(self, prior: AgentExecutorResponse, ctx: WorkflowContext[AgentExecutorResponse]) -> None:
-        """Enable seamless chaining: accept a prior AgentExecutorResponse as input.
-
-        Strategy: treat the prior response's messages as the conversation state and
-        immediately run the agent to produce a new response.
-        """
-        # Replace cache with full conversation if available, else fall back to agent_run_response messages.
-        if prior.full_conversation is not None:
-            self._cache = list(prior.full_conversation)
-        else:
-            self._cache = list(prior.agent_run_response.messages)
-        await self._run_agent_and_emit(ctx)
-
-    @handler
-    async def from_str(self, text: str, ctx: WorkflowContext[AgentExecutorResponse]) -> None:
-        """Accept a raw user prompt string and run the agent (one-shot)."""
-        self._cache = [ChatMessage(role="user", text=text)]  # type: ignore[arg-type]
-        await self._run_agent_and_emit(ctx)
-
-    @handler
-    async def from_message(self, message: ChatMessage, ctx: WorkflowContext[AgentExecutorResponse]) -> None:  # type: ignore[name-defined]
-        """Accept a single ChatMessage as input."""
-        self._cache = [message]
-        await self._run_agent_and_emit(ctx)
-
-    @handler
-    async def from_messages(self, messages: list[ChatMessage], ctx: WorkflowContext[AgentExecutorResponse]) -> None:  # type: ignore[name-defined]
-        """Accept a list of ChatMessage objects as conversation context."""
-        self._cache = list(messages)
-        await self._run_agent_and_emit(ctx)
-
-
-# endregion: Agent Executor
