+++ python/packages/core/agent_framework/_workflows/_sequential.py
@@ -42,10 +42,12 @@ from typing import Any
 
 from agent_framework import AgentProtocol, ChatMessage, Role
 
-from ._checkpoint import CheckpointStorage
-from ._executor import (
+from ._agent_executor import (
     AgentExecutor,
     AgentExecutorResponse,
+)
+from ._checkpoint import CheckpointStorage
+from ._executor import (
     Executor,
     handler,
 )
