+++ python/packages/core/tests/workflow/test_full_conversation.py
@@ -8,22 +8,23 @@ from typing_extensions import Never
 
 from agent_framework import (
     AgentExecutor,
+    AgentExecutorResponse,
     AgentRunResponse,
     AgentRunResponseUpdate,
     AgentThread,
     BaseAgent,
     ChatMessage,
+    Executor,
     Role,
     SequentialBuilder,
     TextContent,
     WorkflowBuilder,
+    WorkflowContext,
     WorkflowOutputEvent,
     WorkflowRunState,
     WorkflowStatusEvent,
     handler,
 )
-from agent_framework._workflows._executor import AgentExecutorResponse, Executor
-from agent_framework._workflows._workflow_context import WorkflowContext
 
 
 class _SimpleAgent(BaseAgent):
